Fix invalid date issues (#53)
* Fix invalid date where seconds > 60

* limit invalid month and day values

* Check valid iso8601 format with isoparse instead of regex
