+++ tests/test_operators.py
@@ -1029,21 +1029,6 @@ class DataframeOperatorTests(TestCase):
             "target": "var2",
             "comparator": ["test", "value"],
         }))
-
-    def test_invalid_date(self):
-        df = pandas.DataFrame.from_dict(
-            {
-                "var1": ['2021', '2021', '2021', '2021', '2099'],
-                "var2": ["2099", "2022", "2034", "90999", "20999"],
-                "var3": ["1997-07", "1997-07-16", "1997-07-16T19:20:30.45+01:00", "1997-07-16T19:20:30+01:00", "1997-07-16T19:20+01:00"], 
-            }
-        )
-        self.assertTrue(DataframeType({"value": df, "column_prefix_map": {"--": "va"}}).invalid_date({"target": "--r1"})
-            .equals(pandas.Series([False, False, False, False, False])))
-        self.assertTrue(DataframeType({"value": df}).invalid_date({"target": "var3"})
-            .equals(pandas.Series([False, False, False, False, False])))
-        self.assertTrue(DataframeType({"value": df}).invalid_date({"target": "var2"})
-            .equals(pandas.Series([False, False, False, True, True])))
     
     def test_date_equal_to(self):
         df = pandas.DataFrame.from_dict(
